move string extension functions to separate file
