+++ library/kmp-tor-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/clientauth/OnionClientAuth.kt
@@ -16,6 +16,7 @@
 package io.matthewnelson.kmp.tor.common.clientauth
 
 import io.matthewnelson.kmp.tor.common.address.OnionAddressV3
+import io.matthewnelson.kmp.tor.common.util.stripString
 import kotlin.jvm.JvmStatic
 import kotlin.jvm.JvmSynthetic
 
@@ -213,19 +214,3 @@ internal inline fun OnionClientAuth.PublicKey.descriptorString(): String =
         .append(':')
         .append(value)
         .toString()
-
-@Suppress("nothing_to_inline")
-private inline fun String.stripString(): String {
-    var limit = length
-
-    // Disregard padding and/or whitespace from end of string
-    while (limit > 0) {
-        val c = this[limit - 1]
-        if (c != '=' && c != '\n' && c != '\r' && c != ' ' && c != '\t') {
-            break
-        }
-        limit--
-    }
-
-    return this.substring(0, limit).trimStart()
-}
