+++ library/kmp-tor-common/src/commonTest/kotlin/io/matthewnelson/kmp/tor/common/address/SchemeUnitTest.kt
@@ -15,6 +15,7 @@
  **/
 package io.matthewnelson.kmp.tor.common.address
 
+import io.matthewnelson.kmp.tor.common.util.separateSchemeFromAddress
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
